fix(server): Redis ZCount 타임스탬프 포맷 "%d"로 통일
ZCount min/max 파라미터에서 "%f" (실수) 포맷을 "%d" (정수) 포맷으로 통일.
stats.go만 "%d"를 사용하고 나머지 8개 파일이 "%f"를 사용하던 불일치 해소.

- handler: news, grid, demo_fire의 now 타입을 float64에서 int64로 변경
- sio: fire_ignite, fire_state, subscribe_viewport, get_fires_compat 동일 적용
- engine: fire_progression의 불필요한 UnixMilli()/1000.0을 Unix()로 단순화
- test: 관련 테스트 파일 포맷 동기화

Closes #87

+++ server/internal/sio/fire_ignite.go
@@ -264,7 +264,7 @@ func registerFireEvent(
 	resolver *geodata.AdminRegionResolver,
 	logger *log.Logger,
 ) (*fireRegistration, error) {
-	now := fmt.Sprintf("%f", float64(time.Now().Unix()))
+	now := fmt.Sprintf("%d", time.Now().Unix())
 	currentGrid := gridID
 	var spreadPath [][2]string
